+++ owncloudApp/src/main/java/com/owncloud/android/presentation/common/BottomSheetFragmentItemView.kt
@@ -2,20 +2,23 @@
  * ownCloud Android client application
  *
  * @author Abel García de Prada
- * Copyright (C) 2020 ownCloud GmbH.
- * <p>
+ * @author Juan Carlos Garrote Gascón
+ *
+ * Copyright (C) 2023 ownCloud GmbH.
+ *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 2,
  * as published by the Free Software Foundation.
- * <p>
+ *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
- * <p>
+ *
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
+
 package com.owncloud.android.presentation.common
 
 import android.content.Context
@@ -36,7 +39,7 @@ class BottomSheetFragmentItemView @JvmOverloads constructor(
     private var _binding: BottomSheetFragmentItemBinding? = null
     private val binding get() = _binding!!
 
-    private var itemIcon: Drawable?
+    var itemIcon: Drawable?
         get() = binding.itemIcon.drawable
         set(value) {
             binding.itemIcon.setImageDrawable(value)
@@ -72,5 +75,9 @@ class BottomSheetFragmentItemView @JvmOverloads constructor(
         binding.itemIcon.setColorFilter(selectedColor)
         binding.itemTitle.setTextColor(selectedColor)
         binding.itemAdditionalIcon.setColorFilter(selectedColor)
+    }
+
+    fun removeDefaultTint() {
+        binding.itemIcon.imageTintList = null
     }
 }