+++ owncloudApp/src/main/java/com/owncloud/android/presentation/files/operations/FileOperation.kt
@@ -2,7 +2,9 @@
  * ownCloud Android client application
  *
  * @author Abel García de Prada
- * Copyright (C) 2021 ownCloud GmbH.
+ * @author Juan Carlos Garrote Gascón
+ *
+ * Copyright (C) 2023 ownCloud GmbH.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 2,
@@ -17,6 +19,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http:></http:>//www.gnu.org/licenses/>.
  */
+
 package com.owncloud.android.presentation.files.operations
 
 import com.owncloud.android.domain.files.model.OCFile
@@ -30,6 +33,7 @@ sealed interface FileOperation {
     data class SynchronizeFileOperation(val fileToSync: OCFile, val accountName: String) : FileOperation
     data class SynchronizeFolderOperation(val folderToSync: OCFile, val accountName: String) : FileOperation
     data class RefreshFolderOperation(val folderToRefresh: OCFile, val shouldSyncContents: Boolean) : FileOperation
+    data class CreateFileWithAppProviderOperation(val accountName: String, val parentContainerId: String, val filename: String) : FileOperation
     data class SetFilesAsAvailableOffline(val filesToUpdate: List<OCFile>) : FileOperation
     data class UnsetFilesAsAvailableOffline(val filesToUpdate: List<OCFile>) : FileOperation
 }