+++ owncloudData/src/main/java/com/owncloud/android/data/appregistry/datasources/implementation/OCRemoteAppRegistryDataSource.kt
@@ -2,6 +2,8 @@
  * ownCloud Android client application
  *
  * @author Abel García de Prada
+ * @author Juan Carlos Garrote Gascón
+ *
  * Copyright (C) 2023 ownCloud GmbH.
  *
  * This program is free software: you can redistribute it and/or modify
@@ -16,6 +18,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
+
 package com.owncloud.android.data.appregistry.datasources.implementation
 
 import com.owncloud.android.data.ClientManager
@@ -44,7 +47,21 @@ class OCRemoteAppRegistryDataSource(
             clientManager.getAppRegistryService(accountName).getUrlToOpenInWeb(
                 openWebEndpoint = openWebEndpoint,
                 fileId = fileId,
-                appName = appName
+                appName = appName,
+            )
+        }
+
+    override fun createFileWithAppProvider(
+        accountName: String,
+        createFileWithAppProviderEndpoint: String,
+        parentContainerId: String,
+        filename: String,
+    ): String =
+        executeRemoteOperation {
+            clientManager.getAppRegistryService(accountName).createFileWithAppProvider(
+                createFileWithAppProviderEndpoint = createFileWithAppProviderEndpoint,
+                parentContainerId = parentContainerId,
+                filename = filename,
             )
         }